// src/app/model/shared/agendamento-servicos.ts
export class AgendamentoServicos {
       id_cliente: number;
       id_locacao: number;
       id_carro: number;
       servico: number;
       id_loja: number;
       hora_agendamento: number;
       data_agendamento: Date;
       km_car_serv: string;
       obs_servico: string;
       macaco_cat: string;
}